Deleted duplicate new trip
