+++ client/components/Home/TripCard.jsx
@@ -11,7 +11,9 @@ const useStyles = makeStyles({
   root: {
     minWidth: 275,
     margin: 10,
-    width: 100
+    height: 300,
+    display: 'flex',
+    flexDirection: 'column'
   },
   bullet: {
     display: 'inline-block',
@@ -26,33 +28,30 @@ const useStyles = makeStyles({
   },
 });
 
-export default function SimpleCard() {
+export default function SimpleCard(props) {
+  const { name, dest, start, end, id } = props;
   const classes = useStyles();
-  const bull = <span className={classes.bullet}>•</span>;
 
   return (
     <Link to='/trip'>
       <Card className={classes.root}>
-        <CardContent>
-          <Typography className={classes.title} color="textSecondary" gutterBottom>
-            Word of the Day
-          </Typography>
-          <Typography variant="h5" component="h2">
-            be{bull}nev{bull}o{bull}lent
-          </Typography>
-          <Typography className={classes.pos} color="textSecondary">
-            adjective
-          </Typography>
-          <Typography variant="body2" component="p">
-            well meaning and kindly.
-            <br />
-            {'"a benevolent smile"'}
-          </Typography>
-        </CardContent>
-        <CardActions>
-          <Button size="small">Learn More</Button>
-        </CardActions>
-      </Card>
+      <CardContent>
+        <Typography variant="h5" component="h2">
+          {name}
+        </Typography>
+        <Typography className={classes.pos} color="textSecondary">
+          {dest}
+        </Typography>
+        <Typography variant="body2" component="p">
+          Start: {start}
+          <br />
+          End: {end}
+        </Typography>
+      </CardContent>
+      {/* <CardActions>
+        <Button size="small" color='secondary' variant='outlined'>View</Button>
+      </CardActions> */}
+    </Card>
     </Link>
   );
 }
