--- client/components/TripCard.jsx
@@ -1,54 +1,0 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-
-const useStyles = makeStyles({
-  root: {
-    minWidth: 275,
-    margin: 10,
-    height: 300,
-    display: 'flex',
-    flexDirection: 'column'
-  },
-  bullet: {
-    display: 'inline-block',
-    margin: '0 2px',
-    transform: 'scale(0.8)',
-  },
-  title: {
-    fontSize: 14,
-  },
-  pos: {
-    marginBottom: 12,
-  },
-});
-
-export default function SimpleCard(props) {
-  const { name, dest, start, end, id } = props;
-  const classes = useStyles();
-
-  return (
-    <Card className={classes.root}>
-      <CardContent>
-        <Typography variant="h5" component="h2">
-          {name}
-        </Typography>
-        <Typography className={classes.pos} color="textSecondary">
-          {dest}
-        </Typography>
-        <Typography variant="body2" component="p">
-          Start: {start}
-          <br />
-          End: {end}
-        </Typography>
-      </CardContent>
-      <CardActions>
-        <Button size="small" color='secondary' variant='outlined'>View</Button>
-      </CardActions>
-    </Card>
-  );
-}
