+++ client/styles/home-container.scss
@@ -31,4 +31,8 @@
     justify-content: center;
   }
 
+  a {
+    text-decoration: none;
+  }
+
 }
